msm: kgsl: Try to replay commands from bad context in  recovery
Replay the all the commands that have not been executed by the GPU
in recovery with preamble turned on for the first command to see
if that succeeds. If it does then we have recovered all contexts.
If that fails then replay commands only from contexts that are
in a good state.

CRs-Fixed: 351680
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
(cherry picked from commit 32240ef6d4ae43ea735aad0b023d2a1305203415)

Conflicts:

	drivers/gpu/msm/adreno.c

Change-Id: Id81e1ba50c83dde37077d2fa2bcbf565e1c604a7
Signed-off-by: Ram Kumar Chakravarthy Chebathini <rcheba@codeaurora.org>

Conflicts:

	drivers/gpu/msm/adreno.c

Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>